event_wrapper: Fix stack overflow for EventWrapper
The PartialEq implementation used a `self == other` statement which
caused the eq function to be infinitely called, hence overflowing the
stack.

This patch fixes this by defining the function more clearly and also
covers the case of resize event being matched only by it's discriminant
tag and not by it's internal field values.